added explanations to key functions and data classes

+++ app/src/main/java/com/example/weatherconditionapplication/data/WeatherUiState.kt
@@ -7,8 +7,7 @@ import com.google.android.gms.maps.model.LatLng
 data class WeatherUiState(
     val weatherInfoList: List<WeatherDataInfo> = listOf(),
     val locationRequired: MutableState<Boolean> = mutableStateOf(false),
-    //val currentLocation: MutableState<LatLng> = mutableStateOf(LatLng(0.0.toDouble(), 0.0.toDouble())),
-    val currentLocation: LatLng = LatLng(0.0.toDouble(), 0.0.toDouble()),
+    val currentLocation: LatLng = LatLng(0.0, 0.0),
     val isShowingWeatherPage: Boolean = true,
     val locationAdress: String = "Istanbul"
 )
